Make next_permutation
O(n) time, O(1) space
265 / 265 test cases passed. Runtime: 88 ms. 5.52th percentile.
My solution seemed optimal, guess not.
I proved on paper that reversing the partial list is same as sorting.

Revelation: others took the inner loop out and made it 2 for loops.
The if condition only needs to be satisfied once, so loop until that
condition is true, then move on.